Made LodDataSSS not be a persistent type
